Factor out api_types.ts into a separate file
This will allow these types to be used internally in the frame, which is a lot less cumbersome than passing around tuples everywhere as we're currently doing. postMessage and IndexedDB will continue to use the tuples.
